Fixed an issue, that caused not enough tanks to be in the interation
While iterating through tanks it is assumed, that all edges have an associated list of valid tanks. This is not the case, when no tanks were generated (i.e. only one valid solution was found for one of the edges).
Thus the program now bails early, if an edge is completed.